added a sources folder
The majority of the contents in the files are probably not what we want to host. But this gives us a starting point. Or an idea of how to tackle it. Maybe we dont need indiviudal pages for each item and a single page that includes everything would be better? Ideas?