Moved navigation to left menu
Removed the navigation tabs at the top so that all navigation is in the left menu. This provides a better overview of the whole doc at a glance, lets you explore the structure more easily, and lets us have more root-level sections if we want. This seems to be the most common approach for technical docs today.

Also changed the home page so that it behaves like any other normal page, including having the left menu.